+++ src/serde.rs
@@ -5,6 +5,7 @@
 use crate::{
 	FmtUtc2k,
 	Utc2k,
+	Weekday,
 };
 use serde::{
 	de,
@@ -18,7 +19,6 @@ use std::fmt;
 
 #[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
 impl<'de> Deserialize<'de> for FmtUtc2k {
-	#[allow(clippy::option_if_let_else)] // That looks bad here.
 	/// # Deserialize.
 	///
 	/// Use the optional `serde` crate feature to enable serialization support.
@@ -35,9 +35,7 @@ impl Serialize for FmtUtc2k {
 	///
 	/// Use the optional `serde` crate feature to enable serialization support.
 	fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-	where S: ser::Serializer {
-		serializer.serialize_str(self.as_str())
-	}
+	where S: ser::Serializer { serializer.serialize_str(self.as_str()) }
 }
 
 
@@ -122,9 +120,70 @@ impl Serialize for Utc2k {
 	///
 	/// Use the optional `serde` crate feature to enable serialization support.
 	fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-	where S: ser::Serializer {
-		serializer.serialize_u32(self.unixtime())
+	where S: ser::Serializer { serializer.serialize_u32(self.unixtime()) }
+}
+
+
+
+#[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
+impl<'de> Deserialize<'de> for Weekday {
+	/// # Deserialize.
+	///
+	/// Use the optional `serde` crate feature to enable serialization support.
+	fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+	where D: de::Deserializer<'de> {
+		struct Visitor;
+
+		macro_rules! invalid {
+			($fn:ident, $ty:ty) => (
+				fn $fn<S>(self, _src: $ty) -> Result<Self::Value, S>
+				where S: de::Error {
+					Err(de::Error::custom(concat!(stringify!($ty), " is unsupported")))
+				}
+			);
+		}
+
+		impl<'de> de::Visitor<'de> for Visitor {
+			type Value = Weekday;
+
+			fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
+				f.write_str("a string representation like 'mon' or 'Monday'")
+			}
+
+			fn visit_str<S>(self, src: &str) -> Result<Self::Value, S>
+			where S: de::Error {
+				Weekday::try_from(src).map_err(|_| de::Error::custom("invalid weekday string"))
+			}
+
+			fn visit_bytes<S>(self, src: &[u8]) -> Result<Self::Value, S>
+			where S: de::Error {
+				Weekday::try_from(src).map_err(|_| de::Error::custom("invalid weekday string"))
+			}
+
+			// Too small to hold an in-range value.
+			invalid!(visit_i32, i32);
+			invalid!(visit_i64, i64);
+			invalid!(visit_u32, u32);
+			invalid!(visit_u64, u64);
+			invalid!(visit_char, char);
+			invalid!(visit_i8, i8);
+			invalid!(visit_i16, i16);
+			invalid!(visit_u8, u8);
+			invalid!(visit_u16, u16);
+		}
+
+		deserializer.deserialize_any(Visitor)
 	}
+}
+
+#[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
+impl Serialize for Weekday {
+	#[inline]
+	/// # Serialize.
+	///
+	/// Use the optional `serde` crate feature to enable serialization support.
+	fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+	where S: ser::Serializer { serializer.serialize_str(self.as_str()) }
 }
 
 
@@ -211,6 +270,21 @@ mod tests {
 			// Deserialization should give us a copy of the original.
 			let d = serde_json::from_str::<Utc2k>(&s).expect("Deserialization failed.");
 			assert_eq!(date, d);
+		}
+	}
+
+	#[test]
+	fn t_serde_weekday() {
+		for day in Weekday::all() {
+			let s = serde_json::to_string(&day).expect("Serialization failed.");
+			assert_eq!(s, format!("\"{}\"", day.as_str()));
+
+			let d = serde_json::from_str::<Weekday>(&s).expect("Deserialization failed.");
+			assert_eq!(d, day);
+
+			let d = serde_json::from_str::<Weekday>(&format!("\"{}\"", day.abbreviation()))
+				.expect("Deserialization (abbr) failed.");
+			assert_eq!(d, day);
 		}
 	}
 }